+++ PA2/src/gpt_matrixsolver.cpp
@@ -13,9 +13,9 @@
 
 
 // Function to perform matrix-vector multiplication O(vec.size()^2)
-std::vector<double> matVecMul(const std::vector<std::vector<double>>& matrix, const std::vector<double>& vec) {
+std::vector<float> matVecMul(const std::vector<std::vector<float>>& matrix, const std::vector<float>& vec) {
     int n = vec.size();
-    std::vector<double> result(n, 0.0);
+    std::vector<float> result(n, 0.0);
     for (int i = 0; i < n; ++i) {
         for (int j = 0; j < n; ++j) {
             result[i] += matrix[i][j] * vec[j];
@@ -25,8 +25,8 @@ std::vector<double> matVecMul(const std::vector<std::vector<double>>& matrix, co
 }
 
 // Function to compute the dot product of two vectors (O(vec1.size()))
-double dotProduct(const std::vector<double>& vec1, const std::vector<double>& vec2) {
-    double result = 0.0;
+float dotProduct(const std::vector<float>& vec1, const std::vector<float>& vec2) {
+    float result = 0.0;
     for (size_t i = 0; i < vec1.size(); ++i) {
         result += vec1[i] * vec2[i];
     }
@@ -34,9 +34,9 @@ double dotProduct(const std::vector<double>& vec1, const std::vector<double>& ve
 }
 
 // Function to add two vectors (O(vec1.size()))
-std::vector<double> vecAdd(const std::vector<double>& vec1, const std::vector<double>& vec2) {
+std::vector<float> vecAdd(const std::vector<float>& vec1, const std::vector<float>& vec2) {
     int n = vec1.size();
-    std::vector<double> result(n);
+    std::vector<float> result(n);
     for (int i = 0; i < n; ++i) {
         result[i] = vec1[i] + vec2[i];
     }
@@ -44,9 +44,9 @@ std::vector<double> vecAdd(const std::vector<double>& vec1, const std::vector<do
 }
 
 // Function to subtract two vectors (O(vec1.size()))
-std::vector<double> vecSub(const std::vector<double>& vec1, const std::vector<double>& vec2) {
+std::vector<float> vecSub(const std::vector<float>& vec1, const std::vector<float>& vec2) {
     int n = vec1.size();
-    std::vector<double> result(n);
+    std::vector<float> result(n);
     for (int i = 0; i < n; ++i) {
         result[i] = vec1[i] - vec2[i];
     }
@@ -54,9 +54,9 @@ std::vector<double> vecSub(const std::vector<double>& vec1, const std::vector<do
 }
 
 // Function to scale a vector by a scalar O(vec.size())
-std::vector<double> vecScale(const std::vector<double>& vec, double scalar) {
+std::vector<float> vecScale(const std::vector<float>& vec, float scalar) {
     int n = vec.size();
-    std::vector<double> result(n);
+    std::vector<float> result(n);
     for (int i = 0; i < n; ++i) {
         result[i] = vec[i] * scalar;
     }
@@ -64,20 +64,20 @@ std::vector<double> vecScale(const std::vector<double>& vec, double scalar) {
 }
 
 // Conjugate Gradient method to solve Q * x + dx = 0
-std::vector<double> conjugateGradient(const std::vector<std::vector<double>>& Q, const std::vector<double>& dx, double tol = 1e-10, int maxIter = 1000) {
+std::vector<float> conjugateGradient(const std::vector<std::vector<float>>& Q, const std::vector<float>& dx, float tol = 1e-10, int maxIter = 1000) {
     int n = dx.size();
-    std::vector<double> x(n, 0.0); //declare output vector
-    std::vector<double> r = vecScale(dx, -1.0); // multiply dx by -1
-    std::vector<double> p = r; 
-    double rsold = dotProduct(r, r);
+    std::vector<float> x(n, 0.0); //declare output vector
+    std::vector<float> r = vecScale(dx, -1.0); // multiply dx by -1
+    std::vector<float> p = r; 
+    float rsold = dotProduct(r, r);
 
     for (int i = 0; i < maxIter; ++i) {
         std::cout << "Running Iteration " << i << " out of " << maxIter << " Total Iterations" << std::endl;
-        std::vector<double> Qp = matVecMul(Q, p);
-        double alpha = rsold / dotProduct(p, Qp);
+        std::vector<float> Qp = matVecMul(Q, p);
+        float alpha = rsold / dotProduct(p, Qp);
         x = vecAdd(x, vecScale(p, alpha));
         r = vecSub(r, vecScale(Qp, alpha));
-        double rsnew = dotProduct(r, r);
+        float rsnew = dotProduct(r, r);
 
         if (sqrt(rsnew) < tol) {
             std::cout << "Finished on Iteration: " << i << std::endl;
